rename auth-image.tsx
