+++ client/src/components/authenticated-image.tsx
